Small fixes and refactoring.
